refactor(suite): general refactoring
