+++ packages/suite/src/types/suite/index.ts
@@ -2,6 +2,7 @@ import type { ThunkDispatch, ThunkAction as TAction } from 'redux-thunk';
 import type { Store as ReduxStore } from 'redux';
 
 import { deviceActions, discoveryActions, transactionsActions } from '@suite-common/wallet-core';
+import { contactsActions } from '@suite-common/contacts';
 import { firmwareActions } from '@suite-common/firmware';
 import { analyticsActions } from '@suite-common/analytics';
 import type { UiEvent, TransportEvent, BlockchainEvent } from '@trezor/connect';
@@ -9,7 +10,6 @@ import { notificationsActions } from '@suite-common/toast-notifications';
 import { messageSystemActions } from '@suite-common/message-system';
 import { deviceAuthenticityActions } from '@suite-common/device-authenticity';
 import type { Route } from '@suite-common/suite-types';
-import { ContactsAction } from '@suite-common/contacts';
 
 import type { RouterAction } from 'src/actions/suite/routerActions';
 import type { AppState } from 'src/reducers/store';
@@ -53,6 +53,7 @@ export type MessageSystemAction = ReturnType<
     (typeof messageSystemActions)[keyof typeof messageSystemActions]
 >;
 type AnalyticsAction = ReturnType<(typeof analyticsActions)[keyof typeof analyticsActions]>;
+type ContactsAction = ReturnType<(typeof contactsActions)[keyof typeof contactsActions]>;
 type FirmwareAction = ReturnType<(typeof firmwareActions)[keyof typeof firmwareActions]>;
 type DeviceAction = ReturnType<(typeof deviceActions)[keyof typeof deviceActions]>;
 type DiscoveryAction = ReturnType<(typeof discoveryActions)[keyof typeof discoveryActions]>;
