+++ suite-common/contacts/src/index.ts
@@ -1,7 +1,6 @@
-export * as contactsActions from './redux/contactsActions';
-export type { ContactsAction } from './redux/contactsActions';
-export { prepareContactsReducer } from './redux/contactsReducer';
+export * from './redux/contactsActions';
+export * from './redux/contactsReducer';
 export * from './redux/contactsSelectors';
-export * as CONTACTS from './constants';
+export * from './constants';
 export * from './types';
-export * from './utils/findContact';
+export * from './utils';
