+++ suite-common/contacts/src/redux/contactsActions.ts
@@ -1,15 +1,18 @@
 import { createAction } from '@reduxjs/toolkit';
 
 import { Contact } from '../types';
-import * as CONTACTS from '../constants';
+import { CONTACTS_MODULE_PREFIX } from '../constants';
 
-export const addContact = createAction(CONTACTS.ADD, (payload: Contact) => ({ payload }));
+export const addContact = createAction(`${CONTACTS_MODULE_PREFIX}/add`, (payload: Contact) => ({
+    payload,
+}));
 
-export const removeContact = createAction(CONTACTS.REMOVE, (payload: Contact) => ({ payload }));
+export const removeContact = createAction(
+    `${CONTACTS_MODULE_PREFIX}/remove`,
+    (payload: Contact) => ({ payload }),
+);
 
 export const contactsActions = {
     addContact,
     removeContact,
 };
-
-export type ContactsAction = ReturnType<typeof addContact> | ReturnType<typeof removeContact>;
