+++ suite-common/contacts/src/utils/index.ts
@@ -1,0 +1,6 @@
+import { TrezorDevice } from '@suite-common/suite-types';
+
+export * from './findContact';
+
+export const getDeviceState = (device: TrezorDevice) =>
+    device.state?.staticSessionId?.split('@')?.[0];
